@@ -2,40 +2,41 @@
   "name": "CASTNXT",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.8.2",
-    "@emotion/styled": "^11.8.1",
-    "@ginkgo-bioworks/react-json-schema-form-builder": "^2.8.0",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/data-grid": "^4.0.0-alpha.37",
-    "@mui/icons-material": "^5.14.14",
-    "@mui/material": "^5.14.14",
-    "@mui/styled-engine-sc": "^5.4.2",
-    "@mui/x-date-pickers": "^5.0.8",
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
+    "@ginkgo-bioworks/react-json-schema-form-builder": "^3.0.1",
+    "@mui/icons-material": "^6.1.3",
+    "@mui/material": "^6.1.3",
+    "@mui/styled-engine-sc": "^6.1.3",
+    "@mui/x-data-grid": "^7.20.0",
+    "@mui/x-date-pickers": "^7.20.0",
     "@rails/webpacker": "5.4.4",
-    "@react-pdf/renderer": "^3.1.9",
-    "@rjsf/core": "^4.1.1",
-    "axios": "^1.6.7",
-    "babel-jest": "^29.5.0",
+    "@react-pdf/renderer": "^4.0.0",
+    "@rjsf/core": "^5.21.2",
+    "@rjsf/material-ui": "^5.21.2",
+    "@rjsf/utils": "^5.21.2",
+    "axios": "^1.7.7",
+    "babel-jest": "^29.7.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
-    "bootstrap": "^5.1.3",
+    "bootstrap": "^5.3.3",
     "cucumber": "^6.0.7",
-    "dayjs": "^1.11.6",
+    "dayjs": "^1.11.13",
     "file-saver": "^2.0.5",
     "html2canvas": "^1.4.1",
-    "jest": "^29.5.0",
-    "jspdf": "^2.5.1",
-    "material-ui-popup-state": "^1.9.3",
-    "npm": "^10.2.0",
+    "jest": "^29.7.0",
+    "jspdf": "^2.5.2",
+    "npm": "^10.9.0",
+    "npm-check-updates": "^17.1.3",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-beautiful-dnd": "^13.1.0",
-    "react-dom": "^17.0.2",
-    "react-multi-select-component": "^4.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-webcam": "^7.1.1",
-    "semantic-ui-react": "^2.1.2",
-    "styled-components": "^5.3.3",
-    "ts-jest": "^29.1.0",
+    "react": "^18.3.1",
+    "react-beautiful-dnd": "^13.1.1",
+    "react-dom": "^18.3.1",
+    "react-multi-select-component": "^4.3.4",
+    "react-router-dom": "^6.27.0",
+    "react-webcam": "^7.2.0",
+    "semantic-ui-react": "^2.1.5",
+    "styled-components": "^6.1.13",
+    "ts-jest": "^29.2.5",
     "webpack": "^4.46.0",
     "webpack-cli": "^3.3.12"
   },
@@ -64,13 +65,13 @@
     "test:file": "jest --coverage"
   },
   "devDependencies": {
-    "@babel/plugin-transform-modules-commonjs": "^7.21.5",
-    "@babel/preset-react": "^7.16.7",
+    "@babel/plugin-transform-modules-commonjs": "^7.25.7",
+    "@babel/preset-react": "^7.25.7",
     "jest-css-modules": "^2.1.0",
-    "jest-environment-jsdom": "^29.3.1",
-    "jest-mock-axios": "^4.7.0-beta",
-    "react-test-renderer": "^17.0.2",
+    "jest-environment-jsdom": "^29.7.0",
+    "jest-mock-axios": "^4.7.3",
+    "react-test-renderer": "^18.3.1",
     "webpack-dev-server": "^3",
-    "yarn": "^1.22.19"
+    "yarn": "^1.22.22"
   }
 }
